♻️💄 Refactor to resolve class too long error
